Crate test_harness

Modules

deposits

run

Structs

Contracts

Setup

TestValidator

Statics

SNAPSHOT_CONTRACTS_1

Uses Chain #1 from the [GANACHE_CONFIG] static to init the contracts

SNAPSHOT_CONTRACTS_1337

ganache-cli setup with deployed contracts using the snapshot directory NOTE: Current the snapshot and test setup use a single Chain.

VALIDATORS